Устанавливает guzzle, подключает карту на странице просмотра задачи
