+++ composer.json
@@ -17,7 +17,8 @@
         "php": ">=5.6.0",
         "yiisoft/yii2": "~2.0.14",
         "yiisoft/yii2-bootstrap4": "~2.0.0",
-        "yiisoft/yii2-swiftmailer": "~2.0.0 || ~2.1.0"
+        "yiisoft/yii2-swiftmailer": "~2.0.0 || ~2.1.0",
+        "guzzlehttp/guzzle": "^7.4"
     },
     "require-dev": {
         "yiisoft/yii2-debug": "~2.1.0",
