+++ config/params.php
@@ -4,4 +4,5 @@ return [
     'adminEmail' => 'admin@example.com',
     'senderEmail' => 'noreply@example.com',
     'senderName' => 'Example.com mailer',
+    'geocoderApiKey' => 'e666f398-c983-4bde-8f14-e3fec900592a',
 ];
